// api/model/dtoArticleResponse.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * BossFi Backend API
 * BossFi区块链后端API服务，支持钱包登录、用户管理、区块链数据同步等功能
 * OpenAPI spec version: 1.0.0
 */
import type { DtoCategoryInfo } from './dtoCategoryInfo';
import type { DtoUserInfo } from './dtoUserInfo';

export interface DtoArticleResponse {
  category?: DtoCategoryInfo;
  category_id?: number;
  comment_count?: number;
  content?: string;
  created_at?: string;
  id?: number;
  images?: string[];
  is_deleted?: boolean;
  like_count?: number;
  title?: string;
  updated_at?: string;
  /** 关联数据 */
  user?: DtoUserInfo;
  user_id?: number;
  view_count?: number;
}